New: lattice with simultaneous Call and Value facts
The change fixes 2 more tests - now 11 tests fail instead
of 13. I also fixed small problems with deep rewriting
of calls and old encapsulation of value facts in singleton
lists. Originally the change was intended to fix intvalue_simple.hn
but it discovered deeper issues with it and fixed other tests.